Fix : 동균코어님 코드리뷰 반영하여 수정 #1

+++ seun/RestAPIStudy/src/main/java/com/seun/restapistudy/domain/Dessert.java
@@ -10,12 +10,12 @@ public class Dessert {
     private Long id;            // 디저트 ID
     private String name;       // 디저트 이름
     private String description; // 디저트 설명
-    private Double price;      // 디저트 가격
-    private Integer calories;   // 디저트 칼로리
+    private Long price;      // 디저트 가격
+    private Long calories;   // 디저트 칼로리
     private String type;       // 디저트 종류
 
     @Builder
-    public Dessert(Long id, String name, String description, Double price, Integer calories, String type) {
+    public Dessert(Long id, String name, String description, Long price, Long calories, String type) {
         this.id = id;
         this.name = name;
         this.description = description;
